chore: Updated uint to uint256

+++ contracts/V8S.sol
@@ -9,67 +9,69 @@ contract V8S {
 
     struct Request {
         bytes data;
-        uint projectId;
+        uint256 projectId;
         bool hasResponse;
         bytes responseData;
     }
 
-    mapping(uint => Project) public projects;
-    mapping(uint => Request) public requests;
+    mapping(uint256 => Project) public projects;
+    mapping(uint256 => Request) public requests;
 
-    uint public nextProjectId = 0;
-    uint public nextRequestId = 0;
+    uint256 public nextProjectId = 0;
+    uint256 public nextRequestId = 0;
 
-    function addProject(string memory url, string memory decodeType) public returns (uint) {
-        uint projectId = nextProjectId++;
+    function addProject(
+        string memory url,
+        string memory decodeType
+    ) public returns (uint256) {
+        uint256 projectId = nextProjectId++;
         projects[projectId] = Project(url, decodeType);
         return projectId;
     }
 
     function addRequest(
-        uint projectId,
+        uint256 projectId,
         bytes memory data
-    ) public returns (uint) {
-        uint requestId = nextRequestId++;
+    ) public returns (uint256) {
+        uint256 requestId = nextRequestId++;
         requests[requestId] = Request(data, projectId, false, "");
         return requestId;
     }
 
-    function addResponse(
-        uint requestId,
-        bytes memory responseData
-    ) public {
+    function addResponse(uint256 requestId, bytes memory responseData) public {
         requests[requestId].responseData = responseData;
         requests[requestId].hasResponse = true;
     }
 
     function getRequest(
-        uint requestId
-    ) public view returns (uint, string memory, bytes memory) {
-        return (requests[requestId].projectId, projects[requests[requestId].projectId].decodeType, requests[requestId].data);
+        uint256 requestId
+    ) public view returns (uint256, string memory, bytes memory) {
+        return (
+            requests[requestId].projectId,
+            projects[requests[requestId].projectId].decodeType,
+            requests[requestId].data
+        );
     }
 
-    function getResponse(
-        uint requestId
-    ) public view returns (bytes memory) {
+    function getResponse(uint256 requestId) public view returns (bytes memory) {
         return requests[requestId].responseData;
     }
 
-    function isResponseExists(uint requestId) public view returns (bool) {
+    function isResponseExists(uint256 requestId) public view returns (bool) {
         return requests[requestId].hasResponse;
     }
-    
-    function getUnrespondedRequests() public view returns (uint[] memory) {
-        uint[] memory unrespondedRequests = new uint[](nextRequestId);
-        uint count = 0;
-        for (uint i = 0; i < nextRequestId; i++) {
+
+    function getUnrespondedRequests() public view returns (uint256[] memory) {
+        uint256[] memory unrespondedRequests = new uint256[](nextRequestId);
+        uint256 count = 0;
+        for (uint256 i = 0; i < nextRequestId; i++) {
             if (!requests[i].hasResponse) {
                 unrespondedRequests[count] = i;
                 count++;
             }
         }
-        uint[] memory result = new uint[](count);
-        for (uint i = 0; i < count; i++) {
+        uint256[] memory result = new uint256[](count);
+        for (uint256 i = 0; i < count; i++) {
             result[i] = unrespondedRequests[i];
         }
         return result;
